feat: split upload and upload with iter
