+++ sn_cli/src/subcommands/files.rs
@@ -11,6 +11,8 @@ mod estimate;
 
 pub(crate) mod download;
 pub(crate) mod upload;
+pub(crate) mod upload_with_iter;
+
 pub(crate) use chunk_manager::ChunkManager;
 
 use clap::Parser;
